+++ message-consumer-app/src/test/java/se/npet/microservices/messageconsumer/service/TemperatureDeviationCalculatorServiceTest.java
@@ -2,6 +2,7 @@ package se.npet.microservices.messageconsumer.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigDecimal;
 import org.junit.jupiter.api.Test;
 import se.npet.microservices.messageconsumer.service.model.Deviation;
 
@@ -11,16 +12,16 @@ class TemperatureDeviationCalculatorServiceTest {
 
   @Test
   void calculateDeviation_HIGH() {
-    assertEquals(Deviation.HIGH, deviationCalculator.calculateDeviation(20.0, 10.0, 5.0));
+    assertEquals(Deviation.HIGH, deviationCalculator.calculateDeviation(BigDecimal.valueOf(20.0), 10.0, 5.0));
   }
 
   @Test
   void calculateDeviation_LOW() {
-    assertEquals(Deviation.LOW, deviationCalculator.calculateDeviation(5.0, 10.0, 5.0));
+    assertEquals(Deviation.LOW, deviationCalculator.calculateDeviation(BigDecimal.valueOf(5.0), 10.0, 5.0));
   }
 
   @Test
   void calculateDeviation_NORMAL() {
-    assertEquals(Deviation.NORMAL, deviationCalculator.calculateDeviation(12.0, 10.0, 5.0));
+    assertEquals(Deviation.NORMAL, deviationCalculator.calculateDeviation(BigDecimal.valueOf(12.0), 10.0, 5.0));
   }
 }
